cleanup mutexes
