+++ lib/src/impl/mutex_impl.dart
@@ -1,0 +1,5 @@
+export 'stub_mutex.dart'
+    // ignore: uri_does_not_exist
+    if (dart.library.io) '../native/native_isolate_mutex.dart'
+    // ignore: uri_does_not_exist
+    if (dart.library.html) '../web/web_mutex.dart';
