+++ lib/src/native/database/native_sqlite_database.dart
@@ -1,9 +1,9 @@
 import 'dart:async';
 import 'dart:isolate';
 
+import 'package:sqlite_async/src/native/native_isolate_mutex.dart';
 import 'package:sqlite_async/src/native/native_sqlite_open_factory.dart';
 
-import '../../../mutex.dart';
 import '../../utils/database_utils.dart';
 import '../../sqlite_connection.dart';
 import '../native_isolate_connection_factory.dart';
