+++ lib/src/web/database/web_sqlite_connection_impl.dart
@@ -1,11 +1,11 @@
 import 'dart:async';
 import 'package:meta/meta.dart';
-import 'package:mutex/mutex.dart';
 import 'package:sqlite_async/src/common/abstract_open_factory.dart';
 
 import 'package:sqlite_async/src/sqlite_connection.dart';
 import 'package:sqlite_async/src/sqlite_queries.dart';
 import 'package:sqlite_async/src/update_notification.dart';
+import 'package:sqlite_async/src/web/web_mutex.dart';
 import 'package:sqlite_async/src/web/web_sqlite_open_factory.dart';
 
 import 'executor/sqlite_executor.dart';
@@ -51,14 +51,14 @@ class WebSqliteConnectionImpl with SqliteQueries implements SqliteConnection {
   Future<T> readLock<T>(Future<T> Function(SqliteReadContext tx) callback,
       {Duration? lockTimeout, String? debugContext}) async {
     await isInitialized;
-    return mutex.protect(() => callback(WebReadContext(executor!)));
+    return mutex.lock(() => callback(WebReadContext(executor!)));
   }
 
   @override
   Future<T> writeLock<T>(Future<T> Function(SqliteWriteContext tx) callback,
       {Duration? lockTimeout, String? debugContext}) async {
     await isInitialized;
-    return mutex.protect(() => callback(WebWriteContext(executor!)));
+    return mutex.lock(() => callback(WebWriteContext(executor!)));
   }
 
   @override
